notes_project/requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.6.4
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
exceptiongroup==1.2.2
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
